water-monitor/requirements.txt
playwright==1.40.0
flask==3.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
